git: move RefName enum from view module
It's no longer used by View API.
